requirement.txt
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
filelock==3.20.0
Flask==3.1.2
fsspec==2025.9.0
hf-xet==1.1.10
huggingface-hub==0.35.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pillow==11.3.0
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
RPi.GPIO==0.7.1
safetensors==0.6.2
setuptools==80.9.0
sympy==1.14.0
timm==1.0.20
tokenizers==0.22.1
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
transformers==4.57.0
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
